Remvoing a warning or two from PrintAction.

+++ source/actions/PrintActions.cc
@@ -1713,7 +1713,7 @@ class cActionPrintTaskProbHistogram : public cAction
           tArray<double> task_prob = genotype->GetTaskProbabilities();     //    get the taks probabilities
           int weight = (m_weighted) ? genotype->GetNumCPUs() : 1;          //    get the proper tally weighting
           for (int k = 0; k < task_prob.GetSize(); k++){                   //    For each task
-            int bin_id = (task_prob[k] < 1.0) ? ceil( ( task_prob[k] * 100 ) / 5 ) : 21;  // find the bin to put it into
+            int bin_id = (task_prob[k] < 1.0) ? (int) ceil( ( task_prob[k] * 100 ) / 5 ) : 21;  // find the bin to put it into
             m_bins(k,bin_id) += weight;                                                   //   ... and tally it
           }
         }
@@ -1724,7 +1724,7 @@ class cActionPrintTaskProbHistogram : public cAction
           int weight = (m_weighted) ? genotype->GetNumOrganisms() : 1;
           tArray<double> task_prob = genotype->GetTaskProbabilities(ctx);
           for (int k = 0; k < task_prob.GetSize(); k++){
-            int bin_id = (task_prob[k] < 1.0) ? ceil( ( task_prob[k] * 100 ) / 5 ) : 21;
+            int bin_id = (task_prob[k] < 1.0) ? (int) ceil( ( task_prob[k] * 100 ) / 5 ) : 21;
             m_bins(k,bin_id) += weight; 
           }
           genotype = genotype->GetNext();  // Next gentoype
